Changes for version 4.
Report syntactic errors by invoking a macro named SYNTACTIC-ERROR,
rather than expanding into a call to a SYNTAX-ERROR procedure, which
in Scheme48 and most other systems is intended to be called from
macro transformer procedures, which we can't do using SYNTAX-RULES.

New iterator LISTING-INTO! similar to LISTING! stores the list in the
cdr of an existing pair.

Change IN-LIST so that it evaluates the successor procedure once at
the beginning of the loop.

Fix typo in definition of MINIMIZING.  It is not, in fact, an alias
for MAXIMIZING.

Update comments throughout the document.