Changes for version 7.
New loop clauses LET and LET-VALUES for simple variable bindings at
each iteration of the loop.  What was called LET is now called WITH.

Eliminate IN-RANGE and IN-NUMBERS in favour of UP-FROM and DOWN-FROM,
which diverge from the IN-... convention for iterators but are
superior in every other respect.

Simplify loop bodies slightly where we can in the output of LOOP.

Fix fence-post error in reverse vector-like iteration.

Update loop clause error reporting in IN-FILE and IN-PORT

Uncomment the example definition of SYNTACTIC-ERROR.