Revert subtle change to semantics of DOWN-FROM.
Fix some indentation mistakes in the process.

The problem lies not in DOWN-FROM but in user termination conditions
and in that there are two edges to the final node, with different
sets of bindings along each one.  Olin Shivers' loop framework, with
bindings-dominate-references scope, solves this problem elegantly.
Lacking BDR scope (which I don't know how to combine with recursion),
it is probably best just to avoid using DOWN-FROM with user-supplied
termination conditions.